requirements.txt
accelerate==1.6.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
av==14.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
diffusers==0.33.1
docker-pycreds==0.4.0
dotenv==0.9.9
einops==0.8.1
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.2
ftfy==6.3.1
fvcore==0.1.5.post20221221
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httptools==0.6.4
huggingface-hub==0.31.2
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
iopath==0.1.10
Jinja2==3.1.6
joblib==1.5.0
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
narwhals==1.39.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.5
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
onnxruntime==1.22.0
openai-clip==1.0.1
opencv-contrib-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
peft==0.15.2
pillow==11.2.1
platformdirs==4.3.8
plotly==6.0.1
portalocker==3.1.1
protobuf==6.30.2
psutil==7.0.0
pycocotools==2.0.8
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
rectified-flow==1.0.1
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
sentry-sdk==2.27.0
setproctitle==1.3.6
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
sympy==1.13.1
tabulate==0.9.0
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.3.30
timm==1.0.15
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.3
triton==3.2.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
wandb==0.19.11
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
yacs==0.1.8
zipp==3.21.0
